chore: split coordinates into multiple files
working on auto calculation
