+++ src/components/pageElements/CornerSubmission/SubmissionReview.jsx
@@ -1,0 +1,205 @@
+import { useContext } from 'react';
+import PropTypes from 'prop-types';
+import { useMutation } from '@tanstack/react-query';
+import { httpsCallable } from 'firebase/functions';
+import { useFunctions } from 'reactfire';
+import { SubmissionContext } from '../../contexts/SubmissionContext.jsx';
+import Wizard from './Wizard.jsx';
+import { keyMap, formatDatum } from '../../helpers/index.mjs';
+
+const Review = () => {
+  const [state, send] = useContext(SubmissionContext);
+
+  const functions = useFunctions();
+  const saveCorner = httpsCallable(functions, 'functions-httpsPostCorner');
+  const { mutate } = useMutation(
+    ['save-corner', state.context.blmPointId],
+    (data) => saveCorner(data),
+    {
+      onSuccess: (response) => {
+        console.log('success', response);
+        state.context = undefined;
+        send({ type: 'NEXT' });
+      },
+      onError: (error) => {
+        console.log('error', error);
+      },
+    }
+  );
+
+  return (
+    <>
+      <div className="grid gap-2">
+        {state.context.type !== 'existing' && (
+          <MetadataReview
+            blmPointId={state.context.blmPointId}
+            {...state.context.metadata}
+          />
+        )}
+        <CoordinateReview
+          datum={state.context.datum}
+          grid={state.context.grid}
+          geographic={state.context.geographic}
+        />
+      </div>
+      <div className="mt-8 flex justify-center">
+        <Wizard
+          back={() => send('BACK')}
+          finish={async () => {
+            await mutate(state.context);
+          }}
+        />
+      </div>
+    </>
+  );
+};
+
+const MetadataReview = ({
+  blmPointId,
+  status,
+  notes,
+  description,
+  accuracy,
+}) => {
+  return (
+    <>
+      <div className="mb-4 flex flex-col text-center">
+        <h2 className="text-xl font-bold uppercase">Corner submission for</h2>
+        <p className="text-lg font-bold">{blmPointId}</p>
+      </div>
+      <div className="flex flex-col">
+        <span className="font-semibold">Monument Status</span>
+        <span className="ml-4">{keyMap.status(status)}</span>
+      </div>
+      <div className="flex flex-col">
+        <span className="font-semibold">Accuracy</span>
+        <span className="ml-4">{keyMap.accuracy(accuracy)}</span>
+      </div>
+      <div className="flex flex-col">
+        <span className="font-semibold">Monument Description</span>
+        <span className="ml-4">{description}</span>
+      </div>
+      <div className="flex flex-col">
+        <span className="font-semibold">General Notes</span>
+        <span className="ml-4">{notes}</span>
+      </div>
+    </>
+  );
+};
+MetadataReview.propTypes = {
+  blmPointId: PropTypes.string,
+  status: PropTypes.string,
+  notes: PropTypes.string,
+  description: PropTypes.string,
+  accuracy: PropTypes.string,
+};
+
+const CoordinateReview = ({ datum, grid, geographic }) => {
+  if (!datum) {
+    return null;
+  }
+
+  const [type] = datum.split('-');
+
+  return (
+    <>
+      <div className="flex justify-between">
+        <span className="font-semibold">Datum</span>
+        <span>{formatDatum(datum)}</span>
+      </div>
+      {type === 'grid' && <GridCoordinateReview grid={grid} />}
+      {type === 'geographic' && (
+        <GeographicCoordinateReview geographic={geographic} />
+      )}
+    </>
+  );
+};
+CoordinateReview.propTypes = {
+  datum: PropTypes.string.isRequired,
+  grid: PropTypes.shape({
+    zone: PropTypes.string,
+    northing: PropTypes.number,
+    easting: PropTypes.number,
+    elevation: PropTypes.number,
+    unit: PropTypes.string,
+  }),
+  geographic: PropTypes.shape({
+    northing: PropTypes.shape({
+      degrees: PropTypes.number,
+      minutes: PropTypes.number,
+      seconds: PropTypes.number,
+    }),
+    easting: PropTypes.shape({
+      degrees: PropTypes.number,
+      minutes: PropTypes.number,
+      seconds: PropTypes.number,
+    }),
+    elevation: PropTypes.number,
+    unit: PropTypes.string,
+  }),
+};
+
+const GridCoordinateReview = ({ grid }) => (
+  <>
+    <div className="flex justify-between">
+      <span className="font-semibold">Zone</span>
+      <span>{keyMap.zone(grid?.zone)}</span>
+    </div>
+    <div className="flex justify-between">
+      <span className="font-semibold">Unit</span>
+      <span>{keyMap.unit(grid?.unit)}</span>
+    </div>
+    <div className="flex justify-between">
+      <span className="font-semibold">Coordinates</span>
+      <span>{`${grid?.northing}, ${grid?.easting}`}</span>
+    </div>
+    {grid?.elevation && (
+      <div className="flex justify-between">
+        <span className="font-semibold">{grid?.verticalDatum} Elevation</span>
+        <span>{grid?.elevation}</span>
+      </div>
+    )}
+  </>
+);
+GridCoordinateReview.propTypes = {
+  grid: PropTypes.shape({
+    zone: PropTypes.string,
+    northing: PropTypes.number,
+    easting: PropTypes.number,
+    elevation: PropTypes.number,
+    unit: PropTypes.string,
+    verticalDatum: PropTypes.string,
+  }),
+};
+
+const GeographicCoordinateReview = ({ geographic }) => (
+  <>
+    <div className="flex justify-between">
+      <span className="font-semibold">Coordinates</span>
+      <span>{`${geographic?.northing?.degrees}° ${geographic?.northing?.minutes}' ${geographic?.northing?.seconds}", `}</span>
+      <span>{`${geographic?.easting?.degrees}° ${geographic?.easting?.minutes}' ${geographic?.easting?.seconds}"`}</span>
+    </div>
+    <div className="flex justify-between">
+      <span className="font-semibold">Ellipsoid Height</span>
+      <span>{`${geographic?.elevation} ${keyMap.unit(geographic?.unit)}`}</span>
+    </div>
+  </>
+);
+GeographicCoordinateReview.propTypes = {
+  geographic: PropTypes.shape({
+    northing: PropTypes.shape({
+      degrees: PropTypes.number,
+      minutes: PropTypes.number,
+      seconds: PropTypes.number,
+    }),
+    easting: PropTypes.shape({
+      degrees: PropTypes.number,
+      minutes: PropTypes.number,
+      seconds: PropTypes.number,
+    }),
+    elevation: PropTypes.number,
+    unit: PropTypes.string,
+  }),
+};
+
+export default Review;
